+++ tbot/machine/channel/subprocess.py
@@ -77,6 +77,7 @@ class SubprocessChannel(channel.Channel):
         maxread = min(1024, max) if max else 1024
         try:
             buf = os.read(self.pty_master, maxread)
+            self._debug_log(buf)
         except BlockingIOError:
             # If we don't get anything, and the timeout hasn't triggered
             # this channel is closed
@@ -87,7 +88,9 @@ class SubprocessChannel(channel.Channel):
                 maxread = min(1024, max - len(buf)) if max else 1024
                 if maxread == 0:
                     break
-                buf += os.read(self.pty_master, maxread)
+                new = os.read(self.pty_master, maxread)
+                buf += new
+                self._debug_log(new)
         except BlockingIOError:
             pass