+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Show any and all output that is received on the channel with `-vvv`
 - `BoardLinuxMachine` now allows the login and password prompts to
   be clobbered
+- `BoardLinuxMachine` login now waits for the shell to respond
 
 
 ## [0.6.2] - 2018-11-22